// src/app/api/models/search-field.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ISearchFieldDataObject } from '../models/i-search-field-data-object';
export interface SearchField {
  autoComplete?: boolean;
  label?: string;
  name?: string;
  type?: string;
  valueList?: Array<ISearchFieldDataObject>;
}